// notification-service/src/domain/entities/NotificationType.enum.ts
export enum NotificationType {
  CONNECTION_REQUEST = "CONNECTION_REQUEST",
  CONNECTION_ACCEPTED = "CONNECTION_ACCEPTED",
  CONNECTION_REJECTED = "CONNECTION_REJECTED",
  EXPERT_APPLICATION_SUBMITTED = "EXPERT_APPLICATION_SUBMITTED",
  EXPERT_APPLICATION_APPROVED = "EXPERT_APPLICATION_APPROVED",
  EXPERT_APPLICATION_REJECTED = "EXPERT_APPLICATION_REJECTED",
}